restructure admin js and remove the types folder
