+++ src/js/admin/components/Tabs/consts.ts
