+++ src/js/admin/globals.d.ts
