+++ src/js/admin/sections/EditFields/Submissions/consts.ts
@@ -1,10 +1,10 @@
-import { ApprovalState, FormID } from 'src/js/types/edit-form.types';
+import type { ApprovalState, FormID } from '@admin/sections/EditFields/consts';
 
 export type SubmissionID = string | number;
 export type SubmissionStatus = 'Confirmed' | 'Declined';
 
 export type SubmissionItem = {
-	id: SubmissionID;
+    id: SubmissionID;
     form_id: FormID;
     fname?: string;
     lname?: string;
@@ -16,21 +16,21 @@ export type SubmissionItem = {
     hide_name?: '0' | '1';
     accept_tos?: '0' | '1';
     submitted_at: string;
-	approval_status: SubmissionStatus;
-	confirmation_token?: string;
+    approval_status: SubmissionStatus;
+    confirmation_token?: string;
 };
 
 export type ChangeAction = 'Confirm' | 'Decline';
 
 
 export type ApprovalStatusProps = {
-	item: SubmissionItem;
-	onStatusChange?: (
-		id: string | number,
-		newStatus: SubmissionStatus,
-		action: ChangeAction
-	) => void;
-	defaultApprovalState?: ApprovalState;
+    item: SubmissionItem;
+    onStatusChange?: (
+        id: string | number,
+        newStatus: SubmissionStatus,
+        action: ChangeAction
+    ) => void;
+    defaultApprovalState?: ApprovalState;
 };
 
 export type Submissions = SubmissionItem[];
